otam project/requirement.txt
customtkinter==5.2.2
darkdetect==0.8.0
ImageIO==2.37.2
lazy_loader==0.4
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
scikit-image==0.25.2
scipy==1.15.3
tifffile==2025.5.10